code cleanup and overhaul
